requirements.txt
pystray
Pillow
sounddevice
numpy
wavio
